new mkUniqueNames' disallows duplicate Pi binders
New version of mkUniqueNames with an extra argument to specify whether
names bound by Pi's are required to be unique.  The default is that
duplicate names are allowed.  If duplicate names are not allowed, an
error is generated if any are found.  This will be used to disallow
duplicate name binders in top-level type signatures.